Changed models to use int for x/y values, added equation_symbol viewset

+++ server/fermat_server/fermatapi/models/equation_symbols.py
@@ -6,6 +6,6 @@ class EquationSymbol(models.Model):
 
 	symbol = models.ForeignKey(Symbol, on_delete=models.DO_NOTHING)
 	equation = models.ForeignKey(Equation, on_delete=models.DO_NOTHING)
-	x = models.FloatField()
-	y = models.FloatField()
+	x = models.IntegerField()
+	y = models.IntegerField()
 	size = models.FloatField()
